# docs/swagger.yaml
basePath: /
definitions:
  domain.File:
    properties:
      name:
        type: string
      size:
        type: integer
    type: object
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      token:
        type: string
    type: object
  transport.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
host: localhost:9999
info:
  contact: {}
  description: Production-ready file server
  title: FileServer API
  version: "1.0"
paths:
  /api/v1/files:
    get:
      description: Get list of uploaded files
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.File'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List files
      tags:
      - files
  /api/v1/files/{name}:
    delete:
      parameters:
      - description: File name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete file
      tags:
      - files
  /api/v1/files/{name}/stream:
    get:
      parameters:
      - description: File name
        in: path
        name: name
        required: true
        type: string
      - description: Force download
        in: query
        name: download
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download or stream file
      tags:
      - files
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Login
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"